% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jasp.t.test.r
\name{bf.t.test}
\alias{bf.t.test}
\title{Bayesian analysis: Bayes Factor T-test}
\usage{
bf.t.test(x = NULL, y = NULL, paired = FALSE, oneSided = FALSE,
  addInformation, ...)
}
\arguments{
\item{x, }{y Vectors of (numeric) data.}

\item{paired}{Logical to specify whether the data in \code{x} and \code{y} are paired data or not (default: FALSE)}

\item{...}{Other parameters passed onto functions that \code{bf.t.test()} calls.}

\item{onesided}{String or logical \code{FALSE} to specify whether a two-sided should be performed (FALSE, default)
or one-sided alternatives (either \code{left} or \code{right})}
}
\description{
The Bayes Factor t-test \code{bf.t.test} relies on code from the JASP Github repository to on the
one hand not have to rely on non-automated JASP usage for doing Bayesian analyses yet on the other hand
obtain the beautiful plots that JASP creates.
}
\seealso{
\code{ttestBF} in the BayesFactor package for information on additional parameters that can be provided
to \code{bf.t.test()}.
}
