% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{lemmens_brewer}
\alias{lemmens_brewer}
\title{Color Brewer Defaults for Theme Lemmens}
\usage{
lemmens_brewer(type = "qual", palette = "Set1", ...)
}
\arguments{
\item{type}{String to indicate type scale: sequential, diverging, or
qualitative (default).}

\item{palette}{One of my favorite palettes: Set1 (default), Paired,
or Dark2.}

\item{...}{Other parameters passed onward to other functions.}
}
\value{
A list that can be added to an existing ggplot object.
}
\description{
The function returns a list of \code{scale_*_brewer} scales with
specific settings that I like together with the \code{theme_lemmens}.
This reduces typing time.
}
