% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\docType{data}
\name{lemmens_solarized}
\alias{lemmens_solarized}
\title{A personalized ggthemes::solarized dark theme.}
\format{An object of class \code{list} of length 5.}
\usage{
lemmens_solarized
}
\description{
a list containing a series of adaptations to make a plot use the
    ggthemes::solarized theme (dark version) with personal adaptations
    particularly regarding strip and legend key background.
}
\keyword{datasets}
