% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{pairsplot}
\alias{pairsplot}
\title{Personalized pairwise correlation plots}
\usage{
pairsplot(mat, ...)
}
\arguments{
\item{mat}{A matrix of numeric (non-categorical) data}
}
\value{
The function returns invisibly.
}
\description{
\code{pairsplot()} creates a pairwase scatter/correlation plots for each
column of the matrix \code{mat} against the other ones. It uses two custom
panel functions that work best with non-categorical data. If you need to do
a smarter way to look at data, consider using ggpairs in the ggally package.
It defaults to calculating the (parametric) Pearson correlation r.
}
\details{
The default paneling functions that \code{pairsplot} uses are listed in the
help file for the \code{pairs()} function.
}
\seealso{
\link{https://ggobi.github.io/ggally/ggpairs.html} for a function that tackles
varying types of data in a smarter way.
\link{http://personality-project.org/r/Rfunc/pairs.panels.R}
}

