% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{lemmens_solarized}
\alias{lemmens_solarized}
\title{A personalized ggthemes::solarized dark theme.}
\usage{
lemmens_solarized(accent = "blue")
}
\arguments{
\item{accent}{String to indicate accent (starting) colour (default: blue).}
}
\description{
a list containing a series of adaptations to make a plot use the
    ggthemes::solarized theme (dark version) with personal adaptations
    particularly regarding strip and legend key background.
}
