% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\docType{data}
\name{theme_lemmens}
\alias{theme_lemmens}
\title{Default Lemmens Theme.}
\format{An object of class \code{list} of length 3.}
\usage{
theme_lemmens
}
\description{
Uses the adapted \code{ggplot2::theme_bw} with colour-brewer styling.
}
\keyword{datasets}
